Organize elements in routes
