+++ src/index.js
@@ -5,14 +5,17 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import App from './App';
 import About from './pages/about';
 import ErrorPage from './pages/notFound';
+import ToDoApp from './pages/ToDoApp';
 
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<App />} />
-        <Route path="about" element={<About />} />
-        <Route path="*" element={<ErrorPage />} />
+        <Route path="/" element={<App />}>
+          <Route index element={<ToDoApp />} />
+          <Route path="about" element={<About />} />
+          <Route path="*" element={<ErrorPage />} />
+        </Route>
       </Routes>
     </BrowserRouter>
   </React.StrictMode>,
